requirements.txt
code-loader-helpers
torchvision==0.16.0
google-cloud==0.34.0
google-auth==2.39.0
google-cloud-storage==3.1.0

py-readability-metrics==1.4.5

transformers==4.46.3
spacy==2.3.9
tensorflow==2.12.0